+++ lib/core/reactive.dart
@@ -1,17 +1,27 @@
 import 'dart:async';
 
+import 'package:rxdart/rxdart.dart';
+
 import './snapshot.dart';
 
+/// A Reactive class that represents a value that can be observed.
+///
+/// The value can be updated, and the change is propagated to its listeners.
 class Reactive<T> {
-  Reactive(this._value);
+  Reactive(this._value) {
+    _controller.add(Snapshot(_value, _value));
+  }
+
   T _value;
   T get value => _value;
 
   final Map<String, List<Reactive<dynamic>>> _children = {};
 
+  /// Returns a stream of snapshots containing the current and new values.
   Stream<Snapshot<T>> get stream => _controller.stream;
-  final _controller = StreamController<Snapshot<T>>.broadcast();
+  final _controller = BehaviorSubject<Snapshot<T>>();
 
+  /// Updates the value and notifies listeners if the value has changed.
   set value(T newValue) {
     if (newValue != _value) {
       _value = newValue;
@@ -38,8 +48,9 @@ class Reactive<T> {
   @override
   int get hashCode => _value.hashCode;
 
-  void addParent(String accessor, Reactive parent) {
-    _setChildren(parent, this, accessor);
+  /// Adds a parent to the current instance and associates the current [Reactive] with an accessor. This is used to create a tree of [Reactive] instances.
+  void addParent(String childAccessor, Reactive parent) {
+    _setChildren(parent, this, childAccessor);
   }
 
   void _setChildren(Reactive parent, Reactive child, String accessor) {
@@ -50,6 +61,7 @@ class Reactive<T> {
     }
   }
 
+  /// Returns the first child of the specified type associated with the given accessor.
   Reactive<ChildType> getOneChild<ChildType>(String accessor) {
     if (_children.containsKey(accessor)) {
       return _children[accessor]?.first as Reactive<ChildType>;
@@ -57,6 +69,7 @@ class Reactive<T> {
     throw Exception('No child found with accessor: $accessor');
   }
 
+  /// Returns all children of the specified type associated with the given accessor.
   List<Reactive<ChildType>> getAllChildren<ChildType>(String accessor) {
     if (_children.containsKey(accessor)) {
       return _children[accessor] as List<Reactive<ChildType>>;
@@ -64,6 +77,7 @@ class Reactive<T> {
     throw Exception('No child found with accessor: $accessor');
   }
 
+  /// Adds an effect function that gets called whenever the value changes.
   void addEffect(Function(T) effect) {
     stream.listen((snapshot) {
       effect(snapshot.newValue);