+++ CHANGELOG.md
@@ -5,3 +5,10 @@
 ## 1.0.1
 
 * Updated Readme and added License
+
+## 1.1.0
+
+* Both Reactive and Computed classes emits initial event on subscription. This adds an additional dependency on the `rxdart` package.
+* Tests are added. Use `dart test` to run these
+* Added class documentation internally
+* Updated readme